changes as per miguels comments moved general into util index.js moved kafka utility to kafka index.js moved breadcrumb from kafka to util version update of error dependency
