+++ package.json
@@ -25,9 +25,8 @@
   ],
   "author": "Modusbox",
   "contributors": [
-    "Dwolla",
-    "Rajiv Mothilal <rajiv.mothilal@modusbox.com>",
-    "Miguel de Barros <miguel.debarros@modusbox.com>"
+    "Modusbox",
+    "Dwolla"
   ],
   "license": "Apache-2.0",
   "bugs": {
@@ -35,7 +34,7 @@
   },
   "homepage": "https://github.com/mojaloop/central-services-shared#readme",
   "dependencies": {
-    "@mojaloop/central-services-error-handling": "7.1.2",
+    "@mojaloop/central-services-error-handling": "7.2.2",
     "@mojaloop/central-services-stream": "6.2.2",
     "axios": "0.19.0",
     "catbox": "10.0.6",
