+++ src/util/kafka/index.js
@@ -33,12 +33,253 @@
  * @module src/handlers/lib/kafka
  */
 
-const Producer = require('./producer')
 const Consumer = require('./consumer')
-const Utility = require('./utility')
+const Mustache = require('mustache')
+const Logger = require('../../logger')
+const Kafka = require('../kafka')
+const Producer = require('./producer')
+const Enum = require('../../enums')
+const StreamingProtocol = require('../streaming/protocol')
+
+/**
+ * @function ParticipantTopicTemplate
+ *
+ * @description Generates a participant topic name from the 3 inputs which are used in the placeholder topic template for participants found in the default.json
+ *
+ * @param {string} template - The template for that needs to be populated
+ * @param {string} participantName - participant name, retrieved from database. Example: 'dfsp1'
+ * @param {string} functionality - the functionality flow. Example: 'transfer'
+ * @param {string} action - the action that applies to the flow. Example: 'prepare'
+ *
+ * @returns {string} - Returns topic name to be created, throws error if failure occurs
+ */
+const participantTopicTemplate = (template, participantName, functionality, action) => {
+  try {
+    return Mustache.render(template, {
+      participantName,
+      functionality,
+      action
+    })
+  } catch (e) {
+    Logger.error(e)
+    throw e
+  }
+}
+
+/**
+ * @function GeneralTopicTemplate
+ *
+ * @description Generates a general topic name from the 2 inputs, which are used in the placeholder general topic template found in the default.json
+ *
+ * @param {string} template - The template for that needs to be populated
+ * @param {string} functionality - the functionality flow. Example: 'transfer'
+ * @param {string} action - the action that applies to the flow. Example: 'prepare'
+ *
+ * @returns {string} - Returns topic name to be created, throws error if failure occurs
+ */
+const generalTopicTemplate = (template, functionality, action) => {
+  try {
+    return Mustache.render(template, { functionality, action })
+  } catch (e) {
+    Logger.error(e)
+    throw e
+  }
+}
+
+/**
+ * @function TransformGeneralTopicName
+ *
+ * @description generalTopicTemplate called which generates a general topic name from the 2 inputs, which are used in the placeholder general topic template found in the default.json
+ *
+ * @param {string} template - The template for that needs to be populated
+ * @param {string} functionality - the functionality flow. Example: 'transfer'
+ * @param {string} action - the action that applies to the flow. Example: 'prepare'
+ *
+ * @returns {string} - Returns topic name to be created, throws error if failure occurs
+ */
+const transformGeneralTopicName = (template, functionality, action) => {
+  try {
+    if (Enum.Kafka.TopicMap[functionality] && Enum.Kafka.TopicMap[functionality][action]) {
+      return generalTopicTemplate(template, Enum.Kafka.TopicMap[functionality][action].functionality, Enum.Kafka.TopicMap[functionality][action].action)
+    }
+    return generalTopicTemplate(template, functionality, action)
+  } catch (err) {
+    Logger.error(err)
+    throw err
+  }
+}
+
+/**
+ * @function TransformGeneralTopicName
+ *
+ * @description participantTopicTemplate called which generates a participant topic name from the 3 inputs, which are used in the placeholder participant topic template found in the default.json
+ *
+ * @param {string} template - The template for that needs to be populated
+ * @param {string} participantName - participant name, retrieved from database. Example: 'dfsp1'
+ * @param {string} functionality - the functionality flow. Example: 'transfer'
+ * @param {string} action - the action that applies to the flow. Example: 'prepare'
+ *
+ * @returns {string} - Returns topic name to be created, throws error if failure occurs
+ */
+const transformAccountToTopicName = (template, participantName, functionality, action) => {
+  try {
+    return participantTopicTemplate(template, participantName, functionality, action)
+  } catch (err) {
+    Logger.error(err)
+    throw err
+  }
+}
+
+/**
+ * @function GetKafkaConfig
+ *
+ * @description participantTopicTemplate called which generates a participant topic name from the 3 inputs, which are used in the placeholder participant topic template found in the default.json
+ *
+ * @param {string} kafkaConfig - This is the Kafka config set in the default.json in you API
+ * @param {string} flow - This is required for the config for the Stream Processing API. Example: 'CONSUMER' ie: note the case of text
+ * @param {string} functionality - the functionality flow. Example: 'TRANSFER' ie: note the case of text
+ * @param {string} action - the action that applies to the flow. Example: 'PREPARE' ie: note the case of text
+ *
+ * @returns {string} - Returns topic name to be created, throws error if failure occurs
+ */
+const getKafkaConfig = (kafkaConfig, flow, functionality, action) => {
+  try {
+    const flowObject = kafkaConfig[flow]
+    const functionalityObject = flowObject[functionality]
+    const actionObject = functionalityObject[action]
+    actionObject.config.logger = Logger
+    return actionObject.config
+  } catch (e) {
+    throw new Error(`No config found for flow='${flow}', functionality='${functionality}', action='${action}'`)
+  }
+}
+
+/**
+ * @function createParticipantTopicConfig
+ *
+ * @param {string} template - The template for that needs to be populated
+ * @param {string} participantName - The participant name
+ * @param {string} functionality - the functionality flow. Example: 'transfer' ie: note the case of text
+ * @param {string} action - the action that applies to the flow. Example: 'prepare' ie: note the case of text
+ * @param {number} partition - optional partition to produce to
+ * @param {*} opaqueKey - optional opaque token, which gets passed along to your delivery reports
+ *
+ * @returns {object} - Returns newly created participant topicConfig
+ */
+const createParticipantTopicConf = (template, participantName, functionality, action, key = null, partition = null, opaqueKey = null) => {
+  return {
+    topicName: transformAccountToTopicName(template, participantName, functionality, action),
+    key,
+    partition,
+    opaqueKey
+  }
+}
+
+/**
+ * @function createGeneralTopicConfig
+ *
+ * @param {string} template - The template for that needs to be populated
+ * @param {string} functionality - the functionality flow. Example: 'transfer' ie: note the case of text
+ * @param {string} action - the action that applies to the flow. Example: 'prepare' ie: note the case of text
+ * @param {string} key - optional key that allows partitioning it occur
+ * @param {number} partition - optional partition to produce to
+ * @param {*} opaqueKey - optional opaque token, which gets passed along to your delivery reports
+ *
+ * @returns {object} - Returns newly created general topicConfig
+ */
+const createGeneralTopicConf = (template, functionality, action, key = null, partition = null, opaqueKey = null) => {
+  return {
+    topicName: transformGeneralTopicName(template, functionality, action),
+    key,
+    partition,
+    opaqueKey
+  }
+}
+
+/**
+ * @function produceGeneralMessage
+ *
+ * @async
+ * @description This is an async method that produces a message against a generated Kafka topic. it is called multiple times
+ *
+ * Kafka.Producer.produceMessage called to persist the message to the configured topic on Kafka
+ * Utility.updateMessageProtocolMetadata called updates the messages metadata
+ * Utility.createGeneralTopicConf called dynamically generates the general topic configuration
+ * Utility.getKafkaConfig called dynamically gets Kafka configuration
+ *
+ * @param {string} defaultKafkaConfig - This is the Kafka config set in the default.json in you API
+ * @param {string} functionality - the functionality flow. Example: 'transfer' ie: note the case of text
+ * @param {string} action - the action that applies to the flow. Example: 'prepare' ie: note the case of text
+ * @param {object} message - a list of messages to consume for the relevant topic
+ * @param {object} state - state of the message being produced
+ * @param {string} key - optional key that allows partitioning it occur
+ *
+ * @returns {object} - Returns a boolean: true if successful, or throws and error if failed
+ */
+const produceGeneralMessage = async (defaultKafkaConfig, functionality, action, message, state, key = null) => {
+  let functionalityMapped = functionality
+  let actionMapped = action
+  if (Enum.Kafka.TopicMap[functionality] && Enum.Kafka.TopicMap[functionality][action]) {
+    functionalityMapped = Enum.Kafka.TopicMap[functionality][action].functionality
+    actionMapped = Enum.Kafka.TopicMap[functionality][action].action
+  }
+  const messageProtocol = StreamingProtocol.updateMessageProtocolMetadata(message, functionality, action, state)
+  const topicConfig = createGeneralTopicConf(defaultKafkaConfig.TOPIC_TEMPLATES.GENERAL_TOPIC_TEMPLATE.TEMPLATE, functionalityMapped, actionMapped, key)
+  const kafkaConfig = getKafkaConfig(defaultKafkaConfig, Enum.Kafka.Config.PRODUCER, functionalityMapped.toUpperCase(), actionMapped.toUpperCase())
+  await Producer.produceMessage(messageProtocol, topicConfig, kafkaConfig)
+  return true
+}
+
+/**
+ * @function produceParticipantMessage
+ *
+ * @async
+ *
+ * @description This is an async method that produces a message against a Kafka generated topic for a specific participant. it is called multiple times
+ *
+ * Kafka.Producer.produceMessage called to persist the message to the configured topic on Kafka
+ * Utility.updateMessageProtocolMetadata called updates the messages metadata
+ * Utility.createParticipantTopicConf called dynamically generates the topic configuration with a participant name
+ * Utility.getKafkaConfig called dynamically gets Kafka configuration
+ *
+ * @param {string} defaultKafkaConfig - This is the Kafka config set in the default.json in you API
+ * @param {string} participantName - the name of the participant for topic creation
+ * @param {string} functionality - the functionality flow. Example: 'transfer' ie: note the case of text
+ * @param {string} action - the action that applies to the flow. Example: 'prepare' ie: note the case of text
+ * @param {object} message - a list of messages to consume for the relevant topic
+ * @param {object} state - state of the message being produced
+ *
+ * @returns {object} - Returns a boolean: true if successful, or throws and error if failed
+ */
+const produceParticipantMessage = async (defaultKafkaConfig, participantName, functionality, action, message, state) => {
+  let functionalityMapped = functionality
+  let actionMapped = action
+  if (Enum.Kafka.TopicMap[functionality] && Enum.Kafka.TopicMap[functionality][action]) {
+    functionalityMapped = Enum.Kafka.TopicMap[functionality][action].functionality
+    actionMapped = Enum.Kafka.TopicMap[functionality][action].action
+  }
+  const messageProtocol = StreamingProtocol.updateMessageProtocolMetadata(message, functionality, action, state)
+  const topicConfig = createParticipantTopicConf(defaultKafkaConfig.TOPIC_TEMPLATES.PARTICIPANT_TOPIC_TEMPLATE.TEMPLATE, participantName, functionalityMapped, actionMapped)
+  const kafkaConfig = getKafkaConfig(defaultKafkaConfig, Enum.Kafka.Config.PRODUCER, functionalityMapped.toUpperCase(), actionMapped.toUpperCase())
+  await Producer.produceMessage(messageProtocol, topicConfig, kafkaConfig)
+  return true
+}
+
+const commitMessageSync = async (kafkaTopic, consumer, message) => {
+  if (!Kafka.Consumer.isConsumerAutoCommitEnabled(kafkaTopic)) {
+    await consumer.commitMessageSync(message)
+  }
+}
 
 module.exports = {
   Producer,
   Consumer,
-  Utility
+  transformAccountToTopicName,
+  transformGeneralTopicName,
+  getKafkaConfig,
+  createParticipantTopicConf,
+  createGeneralTopicConf,
+  produceParticipantMessage,
+  produceGeneralMessage,
+  commitMessageSync
 }
