+++ test/unit/headers/transformer.test.js
@@ -29,7 +29,7 @@ const Test = require('tapes')(require('tape'))
 const Sinon = require('sinon')
 const Transformer = require('../../../src/util').Headers
 const Enum = require('../../../src/enums')
-const Util = require('../../../src/util').General
+const Util = require('../../../src/util')
 
 const headerConfigExample = {
   httpMethod: 'PUT',
